refactor: add unit-web module
