+++ core-commonui/unit-web/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/web/WebViewScreen.kt
@@ -1,4 +1,4 @@
-package com.github.diegoberaldin.raccoonforlemmy.core.commonui.web
+package com.github.diegoberaldin.raccoonforlemmy.unit.web
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
