+++ settings.gradle.kts
@@ -45,3 +45,4 @@ include(":feature-settings")
 include(":core-commonui:unit-ban")
 include(":core-commonui:unit-chat")
 include(":core-commonui:unit-zoomableimage")
+include(":core-commonui:unit-web")
